+++ resources/views/filament/products/compartir-listas-precios-button.blade.php
@@ -1,0 +1,17 @@
+@php
+    $priceLists = \App\Models\PriceList::orderedForDisplay();
+@endphp
+
+@if ($priceLists->isNotEmpty())
+    <x-filament::dropdown placement="bottom-end">
+        <x-slot name="trigger">
+            <x-filament::icon-button
+                icon="heroicon-o-share"
+                label="Compartir lista de precios"
+                tooltip="Compartir lista de precios"
+            />
+        </x-slot>
+
+        @include('filament.price-lists.dropdown-items', ['priceLists' => $priceLists])
+    </x-filament::dropdown>
+@endif