+++ resources/views/filament/topbar/mobile-dashboard-button.blade.php
@@ -6,5 +6,5 @@
     :icon="\Filament\Support\Icons\Heroicon::OutlinedHome"
     icon-size="lg"
     :label="__('Volver al dashboard')"
-    class="lg:hidden"
+    class="ms-2 lg:hidden"
 />